requirements-dev.txt
-r requirements.txt
coverage==7.6.1
django-debug-toolbar==4.4.6
